+++ DatabaseWrapper.SqlServer/DatabaseClient.cs
@@ -34,21 +34,6 @@ namespace DatabaseWrapper.SqlServer
         }
 
         /// <summary>
-        /// Enable or disable logging of queries using the Logger(string msg) method (default: false).
-        /// </summary>
-        public bool LogQueries = false;
-
-        /// <summary>
-        /// Enable or disable logging of query results using the Logger(string msg) method (default: false).
-        /// </summary>
-        public bool LogResults = false;
-
-        /// <summary>
-        /// Method to invoke when sending a log message.
-        /// </summary>
-        public Action<string> Logger = null;
-
-        /// <summary>
         /// Timestamp format.
         /// Default is MM/dd/yyyy hh:mm:ss.fffffff tt.
         /// </summary>
@@ -91,6 +76,22 @@ namespace DatabaseWrapper.SqlServer
             {
                 // https://docs.microsoft.com/en-us/sql/sql-server/maximum-capacity-specifications-for-sql-server
                 return 65536 * 4096;
+            }
+        }
+
+        /// <summary>
+        /// Database settings.
+        /// </summary>
+        public DatabaseSettings Settings
+        {
+            get
+            {
+                return _Settings;
+            }
+            set
+            {
+                if (value == null) throw new ArgumentNullException(nameof(Settings));
+                _Settings = value;
             }
         }
 
@@ -688,8 +689,9 @@ namespace DatabaseWrapper.SqlServer
             if (query.Length > MaxStatementLength) throw new ArgumentException("Query exceeds maximum statement length of " + MaxStatementLength + " characters.");
 
             DataTable result = new DataTable();
-             
-            if (LogQueries && Logger != null) Logger(_Header + "query: " + query);
+
+            if (_Settings.Debug.EnableForQueries && _Settings.Debug.Logger != null)
+                _Settings.Debug.Logger(_Header + "query: " + query);
 
             try
             {
@@ -704,15 +706,15 @@ namespace DatabaseWrapper.SqlServer
                     conn.Close();
                 }
 
-                if (LogResults && Logger != null)
+                if (_Settings.Debug.EnableForResults && _Settings.Debug.Logger != null)
                 {
                     if (result != null)
                     {
-                        Logger(_Header + "result: " + result.Rows.Count + " rows");
+                        _Settings.Debug.Logger(_Header + "result: " + result.Rows.Count + " rows");
                     }
                     else
                     {
-                        Logger(_Header + "result: null");
+                        _Settings.Debug.Logger(_Header + "result: null");
                     }
                 }