+++ Test.Sqlite/Program.cs
@@ -29,11 +29,11 @@ namespace Test.Sqlite
                 string filename = Console.ReadLine();
                 if (String.IsNullOrEmpty(filename)) return;
                 _Settings = new DatabaseSettings(filename);
-                _Database = new DatabaseClient(_Settings);
+                _Settings.Debug.Logger = Logger;
+                _Settings.Debug.EnableForQueries = true;
+                _Settings.Debug.EnableForResults = true;
 
-                _Database.Logger = Logger;
-                _Database.LogQueries = true;
-                _Database.LogResults = true;
+                _Database = new DatabaseClient(_Settings);
 
                 #endregion